requirements.txt
squareup==42.1.0.20250521